# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: FMS Log Nexus API
  description: |
    Centralized logging and monitoring platform API for FMS Log Nexus.
    
    ## Authentication
    
    The API supports two authentication methods:
    - **JWT Bearer Token**: For dashboard users and interactive sessions
    - **API Key**: For service accounts and job integrations
    
    ## Rate Limiting
    
    - Log ingestion: 10,000 requests/minute per API key
    - Dashboard queries: 1,000 requests/minute per user
    
  version: 1.0.0
  contact:
    name: FMS Log Nexus Support
    email: support@company.com
  license:
    name: Proprietary
servers:
  - url: https://lognexus.company.com/api
    description: Production server
  - url: https://localhost:5001/api
    description: Development server
...
security:
  - bearerAuth: []
  - apiKeyAuth: []
paths:
...
  /jobs/executions/{id}:
    get:
      tags: [Jobs]
      summary: Get execution details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Execution details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionDetail'

    put:
      tags: [Jobs]
      summary: Update/complete execution
      security:
        - apiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateExecutionRequest'
      responses:
        '200':
          description: Execution updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionUpdatedResponse'

  # ============================================================================
  # Dashboard Endpoints
  # ============================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token for user authentication

    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for service authentication
...
  schemas:
...
    JobStatus:
      type: string
      enum: [Pending, Running, Completed, Failed, Cancelled, Timeout, Warning]
...
    Execution:
      type: object
      properties:
        id:
          type: string
          format: uuid
        jobId:
          type: string
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        durationMs:
          type: integer
        durationFormatted:
          type: string
        status:
          $ref: '#/components/schemas/JobStatus'
        serverName:
          type: string
        logCount:
          type: integer
        errorCount:
          type: integer
        warningCount:
          type: integer
    ExecutionDetail:
      allOf:
        - $ref: '#/components/schemas/Execution'
        - type: object
          properties:
            jobDisplayName:
              type: string
            triggerType:
              type: string
            triggeredBy:
              type: string
            correlationId:
              type: string
            parameters:
              type: object
            resultSummary:
              type: object
            errorMessage:
              type: string
...
    UpdateExecutionRequest:
      type: object
      required: [status]
      properties:
        status:
          $ref: '#/components/schemas/JobStatus'
        resultSummary:
          type: object
        resultCode:
          type: integer
        errorMessage:
          type: string
        errorCategory:
          type: string
    ExecutionUpdatedResponse:
      type: object
      properties:
        executionId:
          type: string
          format: uuid
        jobId:
          type: string
        status:
          $ref: '#/components/schemas/JobStatus'
        durationMs:
          type: integer
        durationFormatted:
          type: string
...